+++ storage-node/Cargo.toml
@@ -22,3 +22,4 @@ reqwest = "0.12"
 tempfile = "3.10.1"
 rand = "0.8"
 tokio-stream = "0.1"
+sqlite = "0.34"
