+++ storage-node/src/cache/data_store_cache/sqlite/mod.rs
@@ -1,0 +1,9 @@
+use sqlite::Connection;
+use tokio::sync::RwLock;
+
+use crate::cache::policy::DataStoreReplacer;
+
+pub struct SqliteStoreCache<R: DataStoreReplacer> {
+    replacer: RwLock<R>,
+    connection: Connection,
+}
